backend/requirements.txt
fastapi[standard]
pydantic-settings
sqlalchemy
pytest
httpx
python-cas
mysqlclient
python-cas
itsdangerous